feat: logic cleanup
